update: heroku関連の設定ファイルを変更

+++ entrypoint.sh
@@ -1,24 +1,15 @@
 #!/bin/bash
 set -e
 
+# https://stackoverflow.com/questions/55693329/how-to-run-a-rails-app-in-docker-on-heroku
 # Remove a potentially pre-existing server.pid for Rails.
-rm -f /webapp/tmp/pids/server.pid
+if [ -f "/webapp/tmp/pids/server.pid" ]; then
+  rm -f /webapp/tmp/pids/server.pid
+fi
 
-# if [ ! -e "/webapp/tmp" ]; then
-#   mkdir /webapp/tmp
-# fi
+if [ ! -e "${RAILS_ENV}" = "production" ]; then
+  # bundle exec rails assets:precompile
+  bundle exec run bin/webpack
+fi
 
-# if [ ! -e "/webapp/tmp/pids" ]; then
-#   mkdir /webapp/tmp/pids
-# fi
-
-# if [ ! -e "/webapp/tmp/sockets" ]; then
-#   mkdir /webapp/tmp/sockets
-# fi
-
-# if [ ! -e "${RAILS_ENV}" = "production" ]; then
-#   bundle exec rails webpacker:compile
-#   bundle exec rails assets:precompile
-# fi
-# Then exec the container's main process (what's set as CMD in the Dockerfile).
 exec "$@"
